Add changes to version
