+++ src/git-providers/differ.js
@@ -84,3 +84,26 @@ export function getSlides(codes, language) {
       .filter(line => line.middle || line.left || line.right);
   });
 }
+
+export function getChanges(lines) {
+  const changes = [];
+  let currentChange = null;
+  let i = 0;
+  const isNewLine = i => !lines[i].left && lines[i].middle;
+  while (i < lines.length) {
+    if (isNewLine(i)) {
+      if (!currentChange) {
+        currentChange = { start: i };
+      }
+    } else {
+      if (currentChange) {
+        currentChange.end = i - 1;
+        changes.push(currentChange);
+        currentChange = null;
+      }
+    }
+    i++;
+  }
+
+  return changes;
+}
