+++ src/slide.js
@@ -72,39 +72,8 @@ function Slide({ lines, styles, changes }) {
   );
 }
 
-let lastLines = null;
-let lastChanges = null;
-function getChanges(lines) {
-  if (lastLines === lines) {
-    return lastChanges;
-  }
-
-  const changes = [];
-  let currentChange = null;
-  let i = 0;
-  const isNewLine = i => !lines[i].left && lines[i].middle;
-  while (i < lines.length) {
-    if (isNewLine(i)) {
-      if (!currentChange) {
-        currentChange = { start: i };
-      }
-    } else {
-      if (currentChange) {
-        currentChange.end = i - 1;
-        changes.push(currentChange);
-        currentChange = null;
-      }
-    }
-    i++;
-  }
-
-  lastLines = lines;
-  console.log("changes", changes);
-  return changes;
-}
-
-export default function SlideWrapper({ time, lines }) {
+export default function SlideWrapper({ time, version }) {
+  const { lines, changes } = version;
   const styles = animation((time + 1) / 2, lines);
-  const changes = React.useMemo(() => getChanges(lines), [lines]);
   return <Slide lines={lines} styles={styles} changes={changes} />;
 }
